removendo exercicios incompletos

+++ base/050/solver.cpp
@@ -48,10 +48,10 @@ int main()
 
     vector<int> res = matchingStrings(strings, queries);
 
-    for (int i = 0; i < res.size(); i++) {
+    for (int i = 0; i < (int)res.size(); i++) {
         cout << res[i];
 
-        if (i != res.size() - 1) {
+        if (i != (int) res.size() - 1) {
             cout << " ";
         }
     }
